refactored config params
