+++ gpt_researcher/config/variables/base.py
