+++ gpt_researcher/config/variables/retrievers.py
