+++ gpt_researcher/config/variables/test_local.json
